Add files via upload
I gained practical experience in data manipulation, visualization, and interpreting statistical results. This project serves as a strong foundation for further research in understanding the factors that contribute to success in Major League Baseball. I am proud of the work I have accomplished and excited to continue exploring the field of data analysis.